requirements.txt
asgiref==3.2.3
Django==2.2
djangorestframework==3.11.0
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
